Extract basic jdbc tests
