+++ kyuubi-common/src/test/scala/org/apache/kyuubi/operation/BasicJDBCTests.scala
@@ -1,0 +1,348 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.kyuubi.operation
+
+import java.sql.{DriverManager, ResultSet, Statement}
+import java.util.Locale
+
+import org.apache.hive.service.cli.HiveSQLException
+
+import org.apache.kyuubi.{KyuubiFunSuite, Utils}
+import org.apache.kyuubi.operation.meta.ResultSetSchemaConstant._
+
+trait BasicJDBCTests extends KyuubiFunSuite {
+
+  protected val dftSchema = "default"
+  protected val user: String = Utils.currentUser
+  protected def jdbcUrl: String
+
+  protected def withMultipleConnectionJdbcStatement(
+      tableNames: String*)(fs: (Statement => Unit)*): Unit = {
+    val connections = fs.map { _ => DriverManager.getConnection(jdbcUrl, user, "") }
+    val statements = connections.map(_.createStatement())
+
+    try {
+      statements.zip(fs).foreach { case (s, f) => f(s) }
+    } finally {
+      tableNames.foreach { name =>
+        if (name.toUpperCase(Locale.ROOT).startsWith("VIEW")) {
+          statements.head.execute(s"DROP VIEW IF EXISTS $name")
+        } else {
+          statements.head.execute(s"DROP TABLE IF EXISTS $name")
+        }
+      }
+      info("Closing statements")
+      statements.foreach(_.close())
+      info("Closed statements")
+      connections.foreach(_.close())
+      info("Closing connections")
+    }
+  }
+
+  protected def withDatabases(dbNames: String*)(fs: (Statement => Unit)*): Unit = {
+    val connections = fs.map { _ => DriverManager.getConnection(jdbcUrl, user, "") }
+    val statements = connections.map(_.createStatement())
+
+    try {
+      statements.zip(fs).foreach { case (s, f) => f(s) }
+    } finally {
+      dbNames.foreach { name =>
+        statements.head.execute(s"DROP DATABASE IF EXISTS $name")
+      }
+      info("Closing statements")
+      statements.foreach(_.close())
+      info("Closed statements")
+      connections.foreach(_.close())
+      info("Closing connections")
+    }
+  }
+
+  protected def withJdbcStatement(tableNames: String*)(f: Statement => Unit): Unit = {
+    withMultipleConnectionJdbcStatement(tableNames: _*)(f)
+  }
+
+
+  test("get catalogs") {
+    withJdbcStatement() { statement =>
+      val metaData = statement.getConnection.getMetaData
+      val catalogs = metaData.getCatalogs
+      catalogs.next()
+      assert(catalogs.getString(TABLE_CAT) === "spark_catalog")
+      assert(!catalogs.next())
+    }
+  }
+
+  test("get schemas") {
+    def checkResult(rs: ResultSet, dbNames: Seq[String]): Unit = {
+      val expected = dbNames.iterator
+      while(rs.next() || expected.hasNext) {
+        assert(rs.getString("TABLE_SCHEM") === expected.next)
+        assert(rs.getString("TABLE_CATALOG").isEmpty)
+      }
+      // Make sure there are no more elements
+      assert(!rs.next())
+      assert(!expected.hasNext, "All expected schemas should be visited")
+    }
+
+    val dbs = Seq("db1", "db2", "db33", "db44")
+    val dbDflts = Seq("default", "global_temp")
+
+    withDatabases(dbs: _*) { statement =>
+      dbs.foreach(db => statement.execute(s"CREATE DATABASE IF NOT EXISTS $db"))
+      val metaData = statement.getConnection.getMetaData
+
+      Seq("", "%", null, ".*", "_*", "_%", ".%") foreach { pattern =>
+        checkResult(metaData.getSchemas(null, pattern), dbs ++ dbDflts)
+      }
+
+      Seq("db%", "db*") foreach { pattern =>
+        checkResult(metaData.getSchemas(null, pattern), dbs)
+      }
+
+      Seq("db_", "db.") foreach { pattern =>
+        checkResult(metaData.getSchemas(null, pattern), dbs.take(2))
+      }
+
+      checkResult(metaData.getSchemas(null, "db1"), Seq("db1"))
+      checkResult(metaData.getSchemas(null, "db_not_exist"), Seq.empty)
+
+      val e = intercept[HiveSQLException](metaData.getSchemas(null, "*"))
+      assert(e.getCause.getMessage contains "Dangling meta character '*' near index 0\n*\n^")
+    }
+  }
+
+  test("get tables") {
+    val table_test = "table_1_test"
+    val table_external_test = "table_2_test"
+    val view_test = "view_1_test"
+    val view_global_test = "view_2_test"
+    val tables = Seq(table_test, table_external_test, view_test, view_global_test)
+    val schemas = Seq("default", "default", "default", "global_temp")
+    val tableTypes = Seq("MANAGED", "EXTERNAL", "VIEW", "VIEW")
+    withJdbcStatement(view_test, view_global_test, table_test, view_test) { statement =>
+      statement.execute(
+        s"CREATE TABLE IF NOT EXISTS $table_test(key int) USING parquet COMMENT '$table_test'")
+      val loc = Utils.createTempDir()
+      statement.execute(s"CREATE EXTERNAL TABLE IF NOT EXISTS $table_external_test(key int)" +
+        s" COMMENT '$table_external_test' LOCATION '$loc'")
+      statement.execute(s"CREATE VIEW IF NOT EXISTS $view_test COMMENT '$view_test'" +
+        s" AS SELECT * FROM $table_test")
+      statement.execute(s"CREATE GLOBAL TEMP VIEW $view_global_test" +
+        s" COMMENT '$view_global_test' AS SELECT * FROM $table_test")
+
+      val metaData = statement.getConnection.getMetaData
+      val rs1 = metaData.getTables(null, null, null, null)
+      var i = 0
+      while(rs1.next()) {
+        assert(rs1.getString(TABLE_CAT).isEmpty)
+        assert(rs1.getString(TABLE_SCHEM) === schemas(i))
+        assert(rs1.getString(TABLE_NAME) == tables(i))
+        assert(rs1.getString(TABLE_TYPE) == tableTypes(i))
+        assert(rs1.getString(REMARKS) === tables(i).replace(view_global_test, ""))
+        i += 1
+      }
+      assert(i === 4)
+
+      val rs2 = metaData.getTables(null, null, null, Array("VIEW"))
+      i = 2
+      while(rs2.next()) {
+        assert(rs2.getString(TABLE_NAME) == tables(i))
+        i += 1
+      }
+      assert(i === 4)
+
+      val rs3 = metaData.getTables(null, null, "table%", Array("VIEW"))
+      assert(!rs3.next())
+
+      val rs4 = metaData.getTables(null, null, "table%", Array("TABLE"))
+      i = 0
+      while(rs4.next()) {
+        assert(rs4.getString(TABLE_NAME) == tables(i))
+        i += 1
+      }
+      assert(i === 2)
+
+      val rs5 = metaData.getTables(null, "default", "%", Array("VIEW"))
+      i = 2
+      while(rs5.next()) {
+        assert(rs5.getString(TABLE_NAME) == view_test)
+      }
+
+      val e = intercept[HiveSQLException](metaData.getTables(null, "*", null, null))
+      assert(e.getCause.getMessage contains "Dangling meta character '*' near index 0\n*\n^")
+    }
+  }
+
+  test("get type info") {
+    withJdbcStatement() { statement =>
+      val typeInfo = statement.getConnection.getMetaData.getTypeInfo
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "VOID")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.NULL)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "BOOLEAN")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.BOOLEAN)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "TINYINT")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.TINYINT)
+      assert(typeInfo.getInt(PRECISION) === 3)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "SMALLINT")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.SMALLINT)
+      assert(typeInfo.getInt(PRECISION) === 5)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "INTEGER")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.INTEGER)
+      assert(typeInfo.getInt(PRECISION) === 10)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "BIGINT")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.BIGINT)
+      assert(typeInfo.getInt(PRECISION) === 19)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "FLOAT")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.FLOAT)
+      assert(typeInfo.getInt(PRECISION) === 7)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "DOUBLE")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.DOUBLE)
+      assert(typeInfo.getInt(PRECISION) === 15)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "STRING")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.VARCHAR)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "BINARY")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.BINARY)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "DECIMAL")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.DECIMAL)
+      assert(typeInfo.getInt(PRECISION) === 38)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 10)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "DATE")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.DATE)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "TIMESTAMP")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.TIMESTAMP)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 3)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "ARRAY")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.ARRAY)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 0)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "MAP")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.JAVA_OBJECT)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 0)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "STRUCT")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.STRUCT)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 0)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+
+      typeInfo.next()
+      assert(typeInfo.getString(TYPE_NAME) === "INTERVAL")
+      assert(typeInfo.getInt(DATA_TYPE) === java.sql.Types.OTHER)
+      assert(typeInfo.getInt(PRECISION) === 0)
+      assert(typeInfo.getShort(NULLABLE) === 1)
+      assert(!typeInfo.getBoolean(CASE_SENSITIVE))
+      assert(typeInfo.getShort(SEARCHABLE) === 0)
+      assert(typeInfo.getInt(NUM_PREC_RADIX) === 0)
+    }
+  }
+}
