+++ kyuubi-main/src/test/scala/org/apache/kyuubi/operation/KyuubiOperationPerUserSuite.scala
@@ -19,7 +19,9 @@ package org.apache.kyuubi.operation
 
 import org.apache.kyuubi.config.KyuubiConf
 
-class KyuubiOperationPerUserSuite extends KyuubiOperationSuite {
+class KyuubiOperationPerUserSuite extends SharedKyuubiOperationContext with JDBCTests {
+
+  override protected def jdbcUrl: String = getJdbcUrl
 
   override protected val conf: KyuubiConf = {
     KyuubiConf().set(KyuubiConf.ENGINE_SHARED_LEVEL, "user")
