+++ kyuubi-main/src/test/scala/org/apache/kyuubi/operation/SharedKyuubiOperationContext.scala
@@ -17,18 +17,18 @@
 
 package org.apache.kyuubi.operation
 
-import org.apache.kyuubi.Utils
+import org.apache.kyuubi.{KyuubiFunSuite, Utils}
 import org.apache.kyuubi.config.KyuubiConf
-import org.apache.kyuubi.ha.HighAvailabilityConf._
+import org.apache.kyuubi.ha.HighAvailabilityConf.{HA_ZK_ACL_ENABLED, HA_ZK_QUORUM}
 import org.apache.kyuubi.ha.server.EmbeddedZkServer
 import org.apache.kyuubi.server.KyuubiServer
 
-abstract class KyuubiOperationSuite extends JDBCTests {
+trait SharedKyuubiOperationContext extends KyuubiFunSuite {
 
   protected val conf: KyuubiConf
 
   private var zkServer: EmbeddedZkServer = _
-  private var server: KyuubiServer = _
+  protected var server: KyuubiServer = _
 
   override def beforeAll(): Unit = {
     zkServer = new EmbeddedZkServer()
@@ -63,5 +63,5 @@ abstract class KyuubiOperationSuite extends JDBCTests {
     super.afterAll()
   }
 
-  override protected def jdbcUrl: String = s"jdbc:hive2://${server.connectionUrl}/;"
+  protected def getJdbcUrl: String = s"jdbc:hive2://${server.connectionUrl}/;"
 }
